Main

Description

You have an ADT for a hard drive (black blox implementation):

class Disk:
  - int length ()
  - int current_position ()
  - char read()
  - void write(char ch)

and then an ADT for a file system on top of the hard drive:

class FileSystem(Disk d)
  - void create_file(string filename, string contents)
  - Read_handle read_file(string filename)
  - void on_disk_spin()

You can save file metadata in ram. You don't control when or to where the disk spins -- the file system calls the on_disk_spin callback whenever the disk spins.

First, implement create_file and on_disk_spin. e.g. running:

create_file("foo", "bar")
create_file("a", "b")
create_file("baz", "bar")

should give a disk that (eventually, w/ enough spins) has:

filename	contents

foo	bar

a	b

baz	bar

Make sure that your data is set up in a way that allows read_file to be implemented

To extend: You have a mutable struct Read_handle:

Read_handle
 - bool done
 - string contents
 - ...

The done field should be set to true whenever the contents field has all of the contents from the file

Task: implement read_file, and modify on_disk_spin to populate read handles

Documentation

class Disk a where

The provided disk api from the question, translated to haskell

Methods

dlength :: a -> Int

current_position :: a -> IO Int

read :: a -> IO Char

write :: a -> Char -> IO ()

newtype ContiguousData

Store contents of disks and files as Map instead of array for convenience. perf optimization: replace with array

Constructors

ContiguousData

Fields

cData :: Map Int Char

cMap :: (Map Int Char -> Map Int Char) -> ContiguousData -> ContiguousData

Change of basis from Map Int Char to ContiguousData TODO: auto derive this w/ Coercible

data File

File metadata struct

Constructors

File

Fields

name :: String

position :: Int

size :: Int

unwrittenContents :: ContiguousData

type FileName = String

type ReverseLookup = Map Int File

Struct to hold mapping of disk position -> file at that position

data FileSystem

Struct for filesystem metadata

Constructors

FileSystem

Fields

files :: Set File

Set of files on disk

fileNames :: Set FileName

Set of filenames on disk

freverseLookup :: ReverseLookup

Map of disk position -> file

freeIndex :: Int

Index of lowest free position (we don't support delete so no fragmentation)

filesToRead :: Map String (IORef ReadHandle)

Map of filename to mutable read handles

continue :: Applicative f => f ()

Helper function to make code more understandable for imperative people

createFile :: (Monad m, Disk a, MonadReader a m, MonadPlus m) => FileSystem -> String -> String -> m FileSystem

createFile has access to an environment that has access to a disk, and returns a nullable FileSystem

onDiskSpin :: (Monad m, Disk a, MonadReader a m, MonadIO m) => FileSystem -> m FileSystem

onDiskSpin also has access to a disk env, and has access to IO

data ReadHandle

Immutable handle type. Usually passed in an IORef for mutability

Constructors

ReadHandle

Fields

rcontents :: ContiguousData

isReady :: Bool

readFile :: (Monad m, MonadState FileSystem m, MonadIO m) => String -> m (IORef ReadHandle)

readFile has an internal modifyable FileSystem state, has IO, and returns a reference to a ReadHandle

data Simulator a where

We create a DSL to write sequences of disk instructions in

Constructors

Bind :: Simulator a -> (a -> Simulator b) -> Simulator b

Pure :: a -> Simulator a

TryCreateFile :: String -> String -> Simulator ()

SpinTo :: Int -> Simulator ()

TryRead :: String -> Simulator ()

PrintState :: Simulator ()

sampleProgram :: Simulator ()

Sample program using the dsl to create 3 files and try reading them at various times

data ConcreteDisk

Concrete implementation of Disk

Constructors

ConcreteDisk

Fields

cLength :: Int

cCurPos :: IORef Int

cStuff :: IORef ContiguousData

printFs :: FileSystem -> IO ()

Haskell can auto derive printing most data types, but not if the type is impure.

printDisk :: ConcreteDisk -> IO ()

Disk is also impure so it also needs its own print

runSimulation :: (Monad m, MonadReader ConcreteDisk m, MonadState FileSystem m, MonadIO m) => Simulator a -> m a

runSimulation has access to a ConcreteDisk environment, a FileSystem internal state, and IO

startingFileSystem :: FileSystem

mkStartingDisk :: IO ConcreteDisk

main :: IO ()